chore: move provider handler to directory
