+++ main.go
@@ -9,13 +9,10 @@ import (
 	"os"
 	"os/signal"
 	"syscall"
-	"time"
 
 	wrpc "github.com/couchbase-examples/wasmcloud-provider-couchbase/bindings"
-	gocbt "github.com/couchbase/gocb-opentelemetry"
-	"github.com/couchbase/gocb/v2"
-	"go.opentelemetry.io/otel"
-	"go.wasmcloud.dev/provider"
+	"github.com/couchbase-examples/wasmcloud-provider-couchbase/provider"
+	sdk "go.wasmcloud.dev/provider"
 )
 
 func main() {
@@ -40,16 +37,13 @@ func run() error {
 	}()
 
 	// Initialize the provider with callbacks to track linked components
-	providerHandler := Handler{
-		linkedFrom:         make(map[string]map[string]string),
-		clusterConnections: make(map[string]*gocb.Collection),
-	}
+	providerHandler := provider.NewLinkHandler()
 
-	p, err := provider.New(
-		provider.TargetLinkPut(providerHandler.handleNewTargetLink),
-		provider.TargetLinkDel(providerHandler.handleDelTargetLink),
-		provider.HealthCheck(providerHandler.handleHealthCheck),
-		provider.Shutdown(providerHandler.handleShutdown),
+	p, err := sdk.New(
+		sdk.TargetLinkPut(providerHandler.HandleNewTargetLink),
+		sdk.TargetLinkDel(providerHandler.HandleDelTargetLink),
+		sdk.HealthCheck(providerHandler.HandleHealthCheck),
+		sdk.Shutdown(providerHandler.HandleShutdown),
 	)
 	if err != nil {
 		return err
@@ -63,7 +57,7 @@ func run() error {
 	signalCh := make(chan os.Signal, 1)
 
 	// Handle RPC operations
-	stopFunc, err := wrpc.Serve(p.RPCClient, &providerHandler)
+	stopFunc, err := wrpc.Serve(p.RPCClient, providerHandler)
 	if err != nil {
 		p.Shutdown()
 		return err
@@ -87,62 +81,5 @@ func run() error {
 		p.Shutdown()
 		stopFunc()
 	}
-	return nil
-}
-
-// Provider handler functions
-func (h *Handler) handleNewTargetLink(link provider.InterfaceLinkDefinition) error {
-	h.Logger.Info("Handling new target link", "link", link)
-	h.linkedFrom[link.SourceID] = link.TargetConfig
-	couchbaseConnectionArgs, err := validateCouchbaseConfig(link.TargetConfig, link.TargetSecrets)
-	if err != nil {
-		h.Logger.Error("Invalid couchbase target config", "error", err)
-		return err
-	}
-	h.updateCouchbaseCluster(link.SourceID, couchbaseConnectionArgs)
-	return nil
-}
-
-func (h *Handler) updateCouchbaseCluster(sourceId string, connectionArgs CouchbaseConnectionArgs) {
-	// Connect to the cluster
-	cluster, err := gocb.Connect(connectionArgs.ConnectionString, gocb.ClusterOptions{
-		Username: connectionArgs.Username,
-		Password: connectionArgs.Password,
-		Tracer:   gocbt.NewOpenTelemetryRequestTracer(otel.GetTracerProvider()),
-	})
-	if err != nil {
-		h.Logger.Error("unable to connect to couchbase cluster", "error", err)
-		return
-	}
-	var collection *gocb.Collection
-	if connectionArgs.CollectionName != "" && connectionArgs.ScopeName != "" {
-		collection = cluster.Bucket(connectionArgs.BucketName).Scope(connectionArgs.ScopeName).Collection(connectionArgs.CollectionName)
-	} else {
-		collection = cluster.Bucket(connectionArgs.BucketName).DefaultCollection()
-	}
-
-	bucket := cluster.Bucket(connectionArgs.BucketName)
-	if err = bucket.WaitUntilReady(5*time.Second, nil); err != nil {
-		h.Logger.Error("unable to connect to couchbase bucket", "error", err)
-	}
-
-	// Store the connection
-	h.clusterConnections[sourceId] = collection
-}
-
-func (h *Handler) handleDelTargetLink(link provider.InterfaceLinkDefinition) error {
-	h.Logger.Info("Handling del target link", "link", link)
-	delete(h.linkedFrom, link.Target)
-	return nil
-}
-
-func (h *Handler) handleHealthCheck() string {
-	h.Logger.Debug("Handling health check")
-	return "provider healthy"
-}
-
-func (h *Handler) handleShutdown() error {
-	h.Logger.Info("Handling shutdown")
-	// clear(handler.linkedFrom)
 	return nil
 }
