+++ provider/results.go
@@ -1,4 +1,4 @@
-package main
+package provider
 
 import (
 	"github.com/couchbase/gocb/v2"
